base setup
base setup for player interaction with selecting, placing and removing tiles from the board

+++ utils.py
@@ -1,11 +1,12 @@
 def floor(number: float) -> int:
     if round(number) - number > 0:
-        return number - (round(number) - number)
+        return int(number - (round(number) - number))
     else:
         return round(number)
+
 
 def ceil(number: float) -> int:
     if round(number) - number > 0:
         return round(number)
     else:
-        return number - (round(number) - number)
+        return int(number - (round(number) - number))
